% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEERreadr-package.R
\docType{data}
\name{icdo3}
\alias{icdo3}
\title{ICD-O-3 SEER Site/Histology Validation List}
\format{A data frame with 50644 rows and 10 variables
\describe{
\item{site}{}
\item{site_desc}{}
\item{histology}{}
\item{histology_desc}{}
\item{histology_behavior}{}
\item{histology_behavior_desc}{}
}}
\source{
\url{https://seer.cancer.gov/icd-o-3/}
}
\usage{
icdo3
}
\description{
The list of ICD-O-3 SEER Site/Histology codes.
}
\keyword{datasets}
