% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_seer.R
\name{read_seer_col_positions}
\alias{read_seer_col_positions}
\title{Read SEER Columns from SAS file}
\usage{

  read_seer_col_positions(file = "https://seer.cancer.gov/manuals/read.seer.research.nov2017.sas")
}
\arguments{
\item{file}{Path to \code{read.seer.research.sas} file.}
}
\description{
Reads the SEER column specification from the SAS import file.
}
