% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readrSEER-package.R
\docType{package}
\name{readrSEER-package}
\alias{readrSEER}
\alias{readrSEER-package}
\title{readrSEER: Read SEER Fixed Width Data Files}
\description{
Functions to read and process the fixed width data
    files from SEER. Includes functions to recode values based on the
    April 2018 data dictionary.
}
\author{
\strong{Maintainer}: Garrick Aden-Buie \email{Garrick.Aden-Buie@moffitt.org}

Authors:
\itemize{
  \item Vincent Major \email{vincentmajor@nyumc.org}
}

}
\keyword{internal}
