% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode.R
\name{recode_seer}
\alias{recode_seer}
\title{Recode SEER Variables}
\usage{
recode_seer(.data)
}
\arguments{
\item{.data}{The SEER data set}
}
\description{
Recodes SEER variables according to the data dictionary.
Only recodes variables that can be directly looked up.
}
