% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode.R
\name{rename_site_specific}
\alias{rename_site_specific}
\title{Rename Site-Specific Variables}
\usage{
rename_site_specific(.data, name_formatter = snakecase::to_snake_case)
}
\arguments{
\item{.data}{The SEER data set}

\item{name_formatter}{A function that takes will be applied to the
site-specific factor descriptions to convert into valid variable
names. See \code{\link[snakecase:to_snake_case]{snakecase::to_snake_case()}}.}
}
\description{
There are a number of variables that cover site-specific factors.
This function renames those variables according the their labels.
The resulting names are long but descriptive.
}
