% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEERreadr-package.R
\docType{data}
\name{seer_data_dictionary}
\alias{seer_data_dictionary}
\title{SEER Data Dictionary}
\format{A named list of tibbles containing reference codes and their corresponding
labels, as used in the SEER data.}
\source{
\url{https://seer.cancer.gov/data-software/documentation/seerstat/nov2017/TextData.FileDescription.pdf}
}
\usage{
seer_data_dictionary
}
\description{
The SEER Data dictionary is built (semi-manually) from the SEER Research Data
Record Description data dictionary for 1973-2015 SEER incidence data files
available at \url{https://seer.cancer.gov/data-software/documentation/ascii-files.html}.
The current version used in this package is the April 2018 documentation version.
}
\keyword{datasets}
