% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{seer_read_fwf}
\alias{seer_read_fwf}
\title{Read SEER Fixed Width File}
\usage{
seer_read_fwf(file, col_positions = seer_read_col_positions(), ...,
  col_types = readr::cols(.default = readr::col_character()))
}
\arguments{
\item{file}{Path to SEER fixed width file.}

\item{col_positions}{SEER column positions, see \code{\link[=seer_read_col_positions]{seer_read_col_positions()}}.}

\item{...}{Arguments passed on to \code{readr::read_fwf}
\describe{
  \item{locale}{The locale controls defaults that vary from place to place.
The default locale is US-centric (like R), but you can use
\code{\link[=locale]{locale()}} to create your own locale that controls things like
the default time zone, encoding, decimal mark, big mark, and day/month
names.}
  \item{na}{Character vector of strings to use for missing values. Set this
option to \code{character()} to indicate no missing values.}
  \item{comment}{A string used to identify comments. Any text after the
comment characters will be silently ignored.}
  \item{skip}{Number of lines to skip before reading data.}
  \item{n_max}{Maximum number of records to read.}
  \item{guess_max}{Maximum number of records to use for guessing column types.}
  \item{progress}{Display a progress bar? By default it will only display
in an interactive session and not while knitting a document. The display
is updated every 50,000 values and will only display if estimated reading
time is 5 seconds or more. The automatic progress bar can be disabled by
setting option \code{readr.show_progress} to \code{FALSE}.}
}}

\item{col_types}{Specification for column types, default is to return all
as character strings. Use \code{NULL} to rely on \link{readr} or see
\code{\link[readr:read_fwf]{readr::read_fwf()}} for further details.}
}
\description{
Reads a SEER fixed width file using the column dictionary in the provided SAS
file (see \code{\link[=seer_read_col_positions]{seer_read_col_positions()}} for more details). Note that the results
are the raw data reported by SEER with no transformations -- all fields are
imported as character strings by default. You can change this by specifying
the \code{col_types} argument using \link{readr} column specification via
\code{\link[readr:cols]{readr::cols()}}, or you can set \code{col_types = NULL} to let \link{readr} guess the
column type.
}
