% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEERreadr-package.R
\docType{package}
\name{SEERreadr-package}
\alias{SEERreadr}
\alias{SEERreadr-package}
\title{SEERreadr: Read SEER Fixed Width Data Files}
\description{
Functions to read and process the fixed width data
    files from SEER. Includes functions to recode values based on the
    April 2018 data dictionary.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/GerkeLab/SEERreadr}
  \item Report bugs at \url{https://github.com/GerkeLab/SEERreadr/issues}
}

}
\author{
\strong{Maintainer}: Garrick Aden-Buie \email{Garrick.Aden-Buie@moffitt.org}

Authors:
\itemize{
  \item Vincent Major \email{vincentmajor@nyumc.org}
}

}
\keyword{internal}
